feat: step3 구현 머지 및 컨플릭트 해결
