+++ src/test/java/auto/application/MatchLotteryServiceTest.java
@@ -1,5 +1,6 @@
 package auto.application;
 
+import auto.domain.Lotto;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
 
@@ -13,23 +14,23 @@ class MatchLottoServiceTest {
     void getMatchedCountMapTest() {
         // given
         var lotteryNumbersList = List.of(
-                List.of(1, 2, 3, 9, 10, 11),
+                new Lotto(List.of(1, 2, 3, 9, 10, 11)),
 
-                List.of(1, 2, 3, 4, 10, 11),
-                List.of(1, 2, 3, 4, 10, 11),
+                new Lotto(List.of(1, 2, 3, 4, 10, 11)),
+                new Lotto(List.of(1, 2, 3, 4, 10, 11)),
 
-                List.of(1, 2, 3, 4, 5, 11),
-                List.of(1, 2, 3, 4, 5, 11),
-                List.of(1, 2, 3, 4, 5, 11));
+                new Lotto(List.of(1, 2, 3, 4, 5, 11)),
+                new Lotto(List.of(1, 2, 3, 4, 5, 11)),
+                new Lotto(List.of(1, 2, 3, 4, 5, 11)));
         var winningNumbersLastWeek = List.of(1, 2, 3, 4, 5, 6);
         // when
-        Map<Integer, Integer> matchedCountMap = matchLottoService.getMatchedCountMap(lotteryNumbersList, winningNumbersLastWeek);
+        Map<MatchedAmount, Integer> matchedCountMap = matchLottoService.getMatchedCountMap(lotteryNumbersList, winningNumbersLastWeek, 9);
         // then
         SoftAssertions.assertSoftly(assertions -> {
-            assertions.assertThat(matchedCountMap.get(3)).isEqualTo(1);
-            assertions.assertThat(matchedCountMap.get(4)).isEqualTo(2);
-            assertions.assertThat(matchedCountMap.get(5)).isEqualTo(3);
-            assertions.assertThat(matchedCountMap.get(6)).isEqualTo(0);
+            assertions.assertThat(matchedCountMap.get(MatchedAmount.THREE)).isEqualTo(1);
+            assertions.assertThat(matchedCountMap.get(MatchedAmount.FOUR)).isEqualTo(2);
+            assertions.assertThat(matchedCountMap.get(MatchedAmount.FIVE)).isEqualTo(3);
+            assertions.assertThat(matchedCountMap.get(MatchedAmount.SIX)).isEqualTo(0);
         });
     }
 }
